[#138] Move framework drivers to component
