+++ frontend/src/pages/landing/components/index.js
@@ -2,3 +2,4 @@ export { default as GetStarted } from "./GetStarted";
 export { default as ExploreStudies } from "./ExploreStudies";
 export { default as FooterDisclaimer } from "./FooterDisclaimer";
 export { default as Jumbotron } from "./Jumbotron";
+export { default as FrameworkDrivers } from "./FrameworkDrivers";
