+++ frontend/src/pages/landing/components/landingcomp.scss
@@ -28,6 +28,88 @@
   }
 }
 
+#framework-drivers {
+  background-color: #ffffff;
+  width: 100%;
+
+  .ant-collapse > .ant-collapse-item > .ant-collapse-header {
+    padding: 24px 0 16px 0;
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 700;
+    line-height: 28px;
+    letter-spacing: -0.28px;
+    font-family: "TabletGothic";
+    border-bottom: 1px solid $muted-color;
+  }
+
+  .text-wrapper {
+    padding: 3rem 2rem 3rem 3rem;
+    h2 {
+      color: #00403e;
+      font-family: Roc Grotesk;
+      font-size: 30px;
+      font-style: normal;
+      font-weight: 700;
+      line-height: 52px;
+    }
+  }
+
+  .image-wrapper {
+    padding: 3rem 3rem 3rem 2rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    .image {
+      border: 12px solid $bg-color-grey;
+      border-radius: 22px;
+    }
+  }
+
+  .framework-drivers-calculation-wrapper {
+    padding: 24px;
+    height: 275px;
+
+    .driver-label {
+      color: #000;
+      font-family: "RocGrotesk";
+      font-size: 14px;
+      font-style: normal;
+      font-weight: 700;
+      line-height: 24px;
+    }
+
+    .driver-icon-wrapper {
+      border-radius: 100px 0px 100px 100px;
+      background: #01605f;
+      width: 100px;
+      height: 100px;
+      flex-shrink: 0;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+
+      &.price-icon {
+        background: #6de19e;
+      }
+    }
+
+    .math-symbol {
+      color: #45d884;
+      font-family: "TabletGothic";
+      font-size: 52px;
+      font-style: normal;
+      font-weight: 700;
+      line-height: 24px;
+    }
+
+    .driver-calculation-stroke {
+      margin-top: 16px;
+    }
+  }
+}
+
 #get-started {
   padding: 3rem $default-padding-size;
   background: #fcf9f4;
